fix: reduce float size 64 -> 32

+++ src/engine/maths/maths.cr
@@ -2,8 +2,8 @@ module Maths
 
     extend self
 
-    def to_radians(angle : Float64) : Float64
-        angle * Math::PI / 180.0f64
+    def to_radians(angle : Float32) : Float32
+        angle * Math::PI / 180.0_f32
     end
 
 end
